reuse collect paths util
