+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@beforesemicolon/html-plus",
-  "version": "0.3.0",
+  "version": "0.4.0",
   "description": "HTML Template Engine/Language",
   "main": "dist/index.js",
   "files": [
