+++ website/server-test.js
@@ -8,74 +8,62 @@ const learnPage = require('./data/learn.page');
 const {CodeSnippet} = require('./tags/code-snippet');
 const site = require('./data/site.json');
 const packageJSON = require('./../package.json');
+const {collectPaths} = require("./data/collect-paths");
 
 const app = express();
 
-const collectPaths = list => {
-  const paths = new Set();
-  
-  for (let item of list) {
-    if (item.hasOwnProperty('path')) {
-      paths.add(item.path);
-      
-      if (item.list && item.list.length) {
-        Array.from(collectPaths(item.list), (p) => paths.add(p))
-      }
-    }
-  }
-  
-  return paths;
-}
-
 const paths = collectPaths(documentationPage.menu.list);
 
-engine(app, path.resolve(__dirname, './pages'), {
-  staticData: {
-    pages: {
-      documentation: documentationPage,
-      home: homePage,
-      learn: learnPage,
+(async () => {
+  await engine(app, path.resolve(__dirname, './pages'), {
+    staticData: {
+      pages: {
+        documentation: documentationPage,
+        home: homePage,
+        learn: learnPage,
+      },
+      site: {
+        ...site,
+        version: packageJSON.version,
+        license: packageJSON.license
+      },
+      params: {},
+      query: {},
+      path: '/'
     },
-    site: {
-      ...site,
-      version: packageJSON.version,
-      license: packageJSON.license
-    },
-    params: {},
-    query: {},
-    path: '/'
-  },
-  customTags: [
-    CodeSnippet,
-  ],
-  onPageRequest: (req) => {
-    const fullPath = req.path.replace(/(\/|\.html)$/, '');
-    return {
-      path: fullPath
-    }
-  }
-});
-
-app.get('/documentation/:group/:doc?', (req, res, next) => {
-  const ext = path.extname(req.path);
-  
-  if (!ext || ext === '.html') {
-    const fullPath = req.path.replace(/(\/|\.html)$/, '');
-    
-    if (paths.has(fullPath)) {
-      return res.render('documentation', {
+    customTags: [
+      CodeSnippet,
+    ],
+    onPageRequest: (req) => {
+      const fullPath = req.path.replace(/(\/|\.html)$/, '');
+      return {
         path: fullPath
-      });
-    } else {
-      return res.redirect('/404')
+      }
     }
-  }
+  });
   
-  next();
-})
+  app.get('/documentation/:group/:doc?', (req, res, next) => {
+    const ext = path.extname(req.path);
+    
+    if (!ext || ext === '.html') {
+      const fullPath = req.path.replace(/(\/|\.html)$/, '');
+      
+      if (paths.has(fullPath)) {
+        return res.render('documentation', {
+          path: fullPath
+        });
+      } else {
+        return res.redirect('/404')
+      }
+    }
+    
+    next();
+  })
+  
+  const server = http.createServer(app);
+  
+  server.listen(3000, () => {
+    console.log('listening on port http://localhost:3000/');
+  })
+})()
 
-const server = http.createServer(app);
-
-server.listen(3000, () => {
-  console.log('listening on port http://localhost:3000/');
-})
